Move integration files and replace mode selection
